% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_cols.R
\name{percentage_by_group}
\alias{percentage_by_group}
\title{Percentage by Group}
\usage{
percentage_by_group(.data, group_by_cols, conditions = TRUE)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{group_by_cols}{Column names or ids for grouping the rows}

\item{conditions}{Conditions that a row should satisfy to be considered in percentage calculation}
}
\value{
A summarized dataframe with all combinations of group_by_cols, and a new column called "percentage"
}
\description{
Percentage by Group
}
\details{
Summarize the percentage of rows that satisfies given condition(s) by group
}
\examples{
apt_buildings \%>\% filter(year_built >= 1950) \%>\% percentage_by_group('year_built', sprinkler_system == "YES")
apt_buildings \%>\% filter(year_built >= 1950) \%>\% percentage_by_group(c('year_built', 'property_type'), no_of_storeys >= 5)
}
